convert_for_stmt: Move function 'convert_len' to file convert_operators.py test=develop
