--- python/paddle/fluid/dygraph/dygraph_to_static/convert_builtins_func.py
@@ -1,47 +1,0 @@
-# Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import print_function
-
-from paddle.fluid import framework
-from paddle.fluid import core
-from paddle.fluid.layers import nn
-from paddle.fluid.layers import control_flow
-
-
-def convert_len(var):
-    """
-    return variable(length) from shape ops based on var.type
-
-    Note: In addition to some ast transformations, some block-related
-          operations are added in `len` transformation, such as appending
-          `shape_op` in var.block.
-    """
-    if isinstance(var, framework.Variable):
-        if var.type in [
-                core.VarDesc.VarType.LOD_TENSOR,
-                core.VarDesc.VarType.SELECTED_ROWS
-        ]:
-            # Note: Length of var may be known ahead of time in dygraph,
-            # but it probably represents batch size which can be variant.
-            # so we return a variable dynamically inferred from var.shape.
-            return nn.shape(var)[0]
-        elif var.type == core.VarDesc.VarType.LOD_TENSOR_ARRAY:
-            return control_flow.array_length(var)
-        else:
-            raise TypeError(
-                'len(var) only supports LoDTensor/LoDTensorArray/SelectedRows, but received %s.'
-                % type(var))
-    else:
-        return len(var)
